Add centralized handling for calling user callbacks
This change also changes the way outstanding works keeps the runtime
alive.

We now use a counter rather than a single noExitRuntime variable so
that once outstanding works is complete the thread can be shutdown
if needed.

The specific need for this comes the use case of running fetch thread
and wanting to exit the thread once the fetch completes.  Without
this change `pthread_exit` from a fetch callback generated an
`uncaught unwind` message on the console.

Also, if one wants the fetch thread to become joinable there was no way
other than calling `pthread_exit` which has no parallel in the C++
threads world and has issues around calling C++ destructors on the
stack.  With this solution the thread will exit automatically when
returning from the fetch callback.